Introduce nice to have overarching <Headings />
