+++ src/components/headings/index.js
@@ -1,2 +1,3 @@
 /* @flow */
-export { Title, Subtitle } from './headings';
+export { Title, Subtitle } from './titles';
+export { default as Headings } from './Headings';
